+++ lib/govuk_publishing_components/presenters/machine_readable/dataset_schema.rb
@@ -10,6 +10,7 @@ module GovukPublishingComponents
       def structured_data
         # http://schema.org/Dataset
         data = CreativeWorkSchema.new(@page).structured_data
+          .merge(distribution)
           .merge(description)
           .merge(name)
         data["@type"] = "Dataset"
@@ -18,9 +19,20 @@ module GovukPublishingComponents
 
     private
 
-      def description
+      def distribution
+        return {} unless page.attachments
+
         {
-          "description" => (page.body || page.description).slice(0..4999),
+          "distribution" => page.attachments.map { |a| present_attachment(a.with_indifferent_access) }.compact,
+        }
+      end
+
+      def description
+        descr = page.body || page.description
+        return {} unless descr
+
+        {
+          "description" => descr.slice(0..4999),
         }
       end
 
@@ -28,6 +40,27 @@ module GovukPublishingComponents
         {
           "name" => page.title,
         }
+      end
+
+      def present_attachment(attachment)
+        title = attachment[:title]
+        url = attachment[:url]
+        return unless title
+        return unless url
+
+        case attachment[:attachment_type]
+        when "external", "html"
+          {
+            "name" => title,
+            "url" => url,
+          }
+        when "file"
+          {
+            "name" => title,
+            "contentUrl" => url,
+            "encodingFormat" => attachment[:content_type],
+          }.compact
+        end
       end
     end
   end