+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ## Unreleased
 
 * Add heading option to warning component ([PR #1415](https://github.com/alphagov/govuk_publishing_components/pull/1415))
+* Add the "distribution" property to the dataset machine readable metadata ([PR #1416](https://github.com/alphagov/govuk_publishing_components/pull/1416))
+* Fix the dataset machine readable metadata erroring if there is no description set ([PR #1416](https://github.com/alphagov/govuk_publishing_components/pull/1416))
 
 ## 21.36.1